separate baking into hybrid assembly
