+++ DOTS.Animation.Hybrid/Baking.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 59a1fcbdd9d34f25b9e1d01e1d370559
+timeCreated: 1672430769
