+++ DOTS.Animation.Hybrid/Components/AnimationsAuthoring.cs.meta
