+++ DOTS.Animation.Hybrid/Components/BakingComponents.cs
@@ -1,0 +1,18 @@
+﻿using Unity.Entities;
+
+namespace AnimationSystem
+{
+    [TemporaryBakingType]
+    internal struct NeedsBakingTag : IComponentData
+    {
+        
+    }
+    
+    [TemporaryBakingType]
+    internal struct AnimatedEntityBakingInfo : IBufferElementData
+    {
+        public int    ClipIndex;
+        public Entity Entity;
+        public int    IndexInKeyframeArray;
+    }
+}
