+++ DOTS.Animation.Hybrid/Components/SkinnedMeshAuthoring.cs.meta
