+++ DOTS.Animation.Hybrid/TransformBaking.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db3701aa938e15240bfc9a393b2821c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
