+++ DOTS.Animation.Hybrid/TransformBaking/Unity.Rendering.asmref.meta
