+++ DOTS.Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f20829a8b8993843a69fd54492b7e4f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
